Allow oov tokens in vocabulary when invert=False
When invert=False the oov token is unused (and likely never touched by the
developer). We should only error if invert=True and there is a ambiguity
in the inversion lookup.

Also, simplify the vocabulary error checking code, and make the expected
format when specifying a vocabulary including special tokens more explicit.

PiperOrigin-RevId: 397826017